Sprint/2/crud voiture

+++ BO-location/src/main/java/controller/ReservationController.java
@@ -6,10 +6,12 @@ import annotation.Param;
 import annotation.PathVariable;
 import annotation.Post;
 import annotation.RestAPI;
+import exception.UnauthorizedException;
 import model.Reservation;
 import service.ClientService;
 import service.HotelService;
 import service.ReservationService;
+import service.TokenService;
 import util.ModelView;
 
 import java.sql.SQLException;
@@ -23,27 +25,38 @@ public class ReservationController {
     private final ReservationService reservationService = new ReservationService();
     private final ClientService clientService = new ClientService();
     private final HotelService hotelService = new HotelService();
+    private final TokenService tokenService = new TokenService();
 
     @Get("/api/reservations")
     @RestAPI
-    public List<Reservation> readAll() throws SQLException {
+    public List<Reservation> readAll(@Param("token") String token) throws SQLException {
+        if (!tokenService.isValidToken(token)) {
+            throw new UnauthorizedException();
+        }
         return reservationService.readAll();
     }
 
     @Get("/api/reservations/{id}")
     @RestAPI
-    public Reservation read(@PathVariable("id") int id) throws SQLException {
+    public Reservation read(@PathVariable("id") int id, @Param("token") String token) throws SQLException {
+        if (!tokenService.isValidToken(token)) {
+            throw new UnauthorizedException();
+        }
         return reservationService.read(id);
     }
 
     @Post("/api/reservations")
     @RestAPI
     public Reservation create(
+            @Param("token") String token,
             @Param("idClient") int idClient,
             @Param("idHotel") int idHotel,
             @Param("dateHeureArrivee") String dateHeureArrivee,
             @Param("nombrePassager") int nombrePassager
     ) throws SQLException {
+        if (!tokenService.isValidToken(token)) {
+            throw new UnauthorizedException();
+        }
         Reservation reservation = new Reservation(
                 null,
                 clientService.read(idClient),
@@ -59,11 +72,15 @@ public class ReservationController {
     @RestAPI
     public Reservation update(
             @PathVariable("id") int id,
+            @Param("token") String token,
             @Param("idClient") int idClient,
             @Param("idHotel") int idHotel,
             @Param("dateHeureArrivee") String dateHeureArrivee,
             @Param("nombrePassager") int nombrePassager
     ) throws SQLException {
+        if (!tokenService.isValidToken(token)) {
+            throw new UnauthorizedException();
+        }
         Reservation reservation = new Reservation(
                 id,
                 clientService.read(idClient),
@@ -77,7 +94,10 @@ public class ReservationController {
 
     @Post("/api/reservations/{id}/delete")
     @RestAPI
-    public Map<String, Object> delete(@PathVariable("id") int id) throws SQLException {
+    public Map<String, Object> delete(@PathVariable("id") int id, @Param("token") String token) throws SQLException {
+        if (!tokenService.isValidToken(token)) {
+            throw new UnauthorizedException();
+        }
         reservationService.delete(id);
         Map<String, Object> result = new HashMap<>();
         result.put("deleted", true);
@@ -87,7 +107,7 @@ public class ReservationController {
 
     @Get("/reservations/form")
     public ModelView reservationForm() throws SQLException {
-        ModelView view = new ModelView("WEB-INF/reservation-form.jsp");
+        ModelView view = new ModelView("reservations/form");
         view.addObject("clients", clientService.readAll());
         view.addObject("hotels", hotelService.readAll());
         return view;
@@ -109,7 +129,7 @@ public class ReservationController {
         );
         reservationService.create(reservation);
 
-        ModelView view = new ModelView("WEB-INF/reservation-form.jsp");
+        ModelView view = new ModelView("reservations/form");
         view.addObject("clients", clientService.readAll());
         view.addObject("hotels", hotelService.readAll());
         view.addObject("message", "Reservation ajoutee");
